test(e2e): construct cross-chain harness in beforeAll and split l1_to_l2 suite
Move the CrossChainMessagingTest construction out of describe scope and into
beforeAll for the token_bridge, l2_to_l1, and l1_to_l2 suites.

Split l1_to_l2 (~12min in CI) into two suites: l1_to_l2 keeps the
duplicate-message scenario and l1_to_l2_inbox_drift holds the inbox-drift
scenario, each ~6-7min. The shared L1 to L2 message helpers are extracted to
message_test_helpers.ts.
